refactor: migrate to Nuxt path aliases
Replaces `@@/` root path aliases with Nuxt-native `#shared/` and `#layers/` aliases throughout the codebase for better consistency with Nuxt conventions.

Moves schema files from root `/schemas` directory to a shared location accessible via `#shared/schemas/` alias, improving module organization and eliminating the need for custom TypeScript configuration.

Updates documentation to reflect simplified import patterns and removes outdated path alias references. Cleans up verbose test command examples and unnecessary folder structure details.
