+++ app/types/index.ts
@@ -1,2 +1,2 @@
+export * from '#layers/dashboard/app/types'
 export * from './link'
-export * from '@@/layers/dashboard/app/types'
