+++ layers/dashboard/app/components/dashboard/migrate/Import.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import type { ImportData } from '@@/schemas/import'
-import { ImportDataSchema } from '@@/schemas/import'
+import type { ImportData } from '#shared/schemas/import'
+import { ImportDataSchema } from '#shared/schemas/import'
 import { AlertCircle, CheckCircle, Download, SkipForward, Upload, XCircle } from 'lucide-vue-next'
 import { toast } from 'vue-sonner'
