+++ server/api/logs/locations.get.ts
@@ -1,5 +1,5 @@
 import type { H3Event } from 'h3'
-import { QuerySchema } from '@@/schemas/query'
+import { QuerySchema } from '#shared/schemas/query'
 
 const { select, and, notEq } = SqlBricks
