+++ server/api/stats/views.get.ts
@@ -1,5 +1,5 @@
 import type { H3Event } from 'h3'
-import { QuerySchema } from '@@/schemas/query'
+import { QuerySchema } from '#shared/schemas/query'
 import { z } from 'zod'
 
 const { select } = SqlBricks
