+++ server/routes/_assets/[...key].get.ts
@@ -1,4 +1,4 @@
-import { LinkSchema } from '@@/schemas/link'
+import { LinkSchema } from '#shared/schemas/link'
 
 const slugValidator = LinkSchema.shape.slug
