+++ shared/schemas/link.ts
