+++ shared/schemas/query.ts
